Going Back in Time and Publish 1.0.0 Without Go Module Support
This commit does something very dirty but goes back and reverts
the slackscot version to 1.0.0 with all the recent improvements
and changes in order to allow a final version fully compatible with
prior versions of go not using go modules. I'm 🤞 that
youppi is still the only dependent slackscot instance at this
point. If not, I apologize and volunteer to fix your downstream
project 😬.