Have Message Updates Use the Original Message's Timestamp
Since the Timestamp is part of a message identification, slackscot
now uses the timestamp of the original message when processing message
updates. This is useful because plugins can use this to know when looking
at a message that it's the same as one previously seen.

One example of this is the fingerquoter that uses the timestamp as a seed to
random generation to decide on a match/trigger. Previously, a message update
would result in a different seed which would cause the previous reaction to be
deleted. Now fingerquoter can remain coherent as it sees the same timestamp/seed
which results in the same match/non-match. The reaction can change depending on the
message content changing but that seems consistent and reasonable too.